moved auth helpers to public/js/authentication.js
