+++ routes/auth.js
@@ -8,6 +8,7 @@
 import { Router } from "express";
 import { config } from "dotenv";
 import axios from "axios";
+import authentication from "../public/js/authentication.js";
 
 config(); // load environment vars from .env file
 
@@ -17,16 +18,7 @@ const router = Router();
 router
     .route("/spotify")
     .get((req, res) => {
-        const scope = "playlist-read-private user-read-private user-library-read"; // TODO: refine scopes to be minimal
-        const userAuthQuery = {
-            response_type: "code",
-            redirect_uri: process.env.SPOTIFY_REDIRECT,
-            client_id: process.env.SPOTIFY_CLIENT,
-            scope: scope
-        };
-
-        const authURL = "https://accounts.spotify.com/authorize?" + // construct http query
-            Object.keys(userAuthQuery).map((currKey) => `${currKey}=${userAuthQuery[currKey]}`).join("&");
+        const authURL = authentication.getSpotifyAuthorizationURL();
 
         return res.redirect(authURL); 
     });
